+++ demo/src/app/components/+button/basic/button-basic.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-demo-button-basic',
+    templateUrl: './button-basic.component.html'
+})
+export class DemoButtonBasicComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit(): void {}
+}