+++ demo/src/app/components/+tooltip/api-parameters.ts
@@ -46,5 +46,17 @@ export const apiParameters = [
         description: '传入template时，注入给template的数据',
         type: 'any',
         default: 'null'
+    },
+    {
+        property: 'thyTooltipOffset',
+        description: '偏移量',
+        type: 'number',
+        default: '4'
+    },
+    {
+        property: 'thyTooltipPin',
+        description: 'hover触发方式下 鼠标移入Tooltip是否固定Tooltip',
+        type: 'boolean',
+        default: false
     }
 ];