+++ demo/src/app/global/+reboot/editable/editable.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-demo-reboot-editable',
+    templateUrl: './editable.component.html'
+})
+export class DemoRebootEditableComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit(): void {}
+}