+++ demo/src/app/global/+reboot/text/text.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-demo-reboot-text',
+    templateUrl: './text.component.html'
+})
+export class DemoRebootTextComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit(): void {}
+}