+++ demo/src/app/global/+variables/module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { DemoVariablesSectionComponent } from './variables.component';
+import { SharedModule } from '../../shared.module';
+
+@NgModule({
+    declarations: [DemoVariablesSectionComponent],
+    imports: [SharedModule],
+    exports: [],
+    providers: []
+})
+export class DemoVariablesModule {}